current status on testing
